requirements.txt
spotipy
pandas
numpy
matplotlib
seaborn
scikit-learn
streamlit
python-dotenv